refactor: moves parts into utils and types
