+++ src/App.tsx
@@ -1,44 +1,18 @@
 import { useEffect, useState } from 'react';
 import Chooser from './Chooser';
 import Footer from './Footer';
-import { Board, GameState, Player } from './types';
-import { generateNewBoard, availableMoves, winCheck, drawCheck } from './utils';
-
-interface GameData {
-    board: Board;
-    gameStatus: GameState;
-}
-
-const getNewGameStatus = (board: Board, currentToken: 'X' | 'O', currentPlayer: Player) => {
-    if (winCheck(board, currentToken)) return currentPlayer === Player.HUMAN ? GameState.WIN : GameState.LOSE;
-    if (drawCheck(board)) return GameState.DRAW;
-    return currentPlayer === Player.HUMAN ? GameState.AI_TURN : GameState.HUMAN_TURN;
-};
-
-interface UpdatedGameDataOptions {
-    currentBoard: Board;
-    currentToken: 'X' | 'O';
-    currentPlayer: Player;
-    squareIndex: number;
-}
-
-const getUpdatedGameData = ({ currentBoard, currentToken, currentPlayer, squareIndex }: UpdatedGameDataOptions) => {
-    const board = currentBoard.map((square, index) =>
-        squareIndex === index ? { ...square, currentPlayer: currentToken } : square,
-    );
-    const gameStatus = getNewGameStatus(board, currentToken, currentPlayer);
-    return { board, gameStatus };
-};
+import { GameData, GameState, Player, Token } from './types';
+import { generateNewBoard, availableMoves, getUpdatedGameData } from './utils';
 
 const App = () => {
-    const [humanToken, setHumanToken] = useState<'X' | 'O'>('X');
+    const [humanToken, setHumanToken] = useState<Token>(Token.X);
     const [gameData, setGameData] = useState<GameData>({
         board: generateNewBoard(),
         gameStatus: GameState.CHOOSING_PLAYER,
     });
 
     const { board, gameStatus } = gameData;
-    const aiToken = humanToken === 'X' ? 'O' : 'X';
+    const aiToken = humanToken === Token.X ? Token.O : Token.X;
 
     const gameInit = () => {
         setGameData((currentGameData) => ({
@@ -53,10 +27,10 @@ const App = () => {
         gameInit();
     };
 
-    // perform player move
     const handleSquarePress = (event: any) => {
         const squareIndex = parseInt(event.target.id, 10);
         if (gameData.gameStatus === GameState.HUMAN_TURN && gameData.board[squareIndex].currentPlayer === null) {
+            // perform player move
             setGameData((currentGameData) => {
                 return getUpdatedGameData({
                     currentBoard: currentGameData.board,
