+++ src/Chooser.tsx
@@ -1,3 +1,5 @@
+import { Token } from './types';
+
 interface ChooserProps {
     onChooserClick: (event: any) => void;
 }
@@ -5,10 +7,10 @@ interface ChooserProps {
 const Chooser = ({ onChooserClick }: ChooserProps) => (
     <div className="chooser">
         <h3>choose your token</h3>
-        <div className="chooser-button" id="X" onClick={onChooserClick}>
+        <div className="chooser-button" id={Token.X} onClick={onChooserClick}>
             X
         </div>
-        <div className="chooser-button" id="O" onClick={onChooserClick}>
+        <div className="chooser-button" id={Token.O} onClick={onChooserClick}>
             O
         </div>
     </div>
