+++ src/utils.ts
@@ -1,4 +1,4 @@
-import { Board } from './types';
+import { Board, GameState, Player, Token, UpdatedGameDataOptions } from './types';
 
 const TERMINAL_STATES = [
     [0, 1, 2],
@@ -102,3 +102,22 @@ export function minimax(tempBoard: Board, turn: 'X' | 'O', depth: number): MoveO
 export function generateNewBoard(): Board {
     return Array.from({ length: 9 }, (_, index) => ({ id: index, currentPlayer: null }));
 }
+
+const getNewGameStatus = (board: Board, currentToken: Token, currentPlayer: Player) => {
+    if (winCheck(board, currentToken)) return currentPlayer === Player.HUMAN ? GameState.WIN : GameState.LOSE;
+    if (drawCheck(board)) return GameState.DRAW;
+    return currentPlayer === Player.HUMAN ? GameState.AI_TURN : GameState.HUMAN_TURN;
+};
+
+export const getUpdatedGameData = ({
+    currentBoard,
+    currentToken,
+    currentPlayer,
+    squareIndex,
+}: UpdatedGameDataOptions) => {
+    const board = currentBoard.map((square, index) =>
+        squareIndex === index ? { ...square, currentPlayer: currentToken } : square,
+    );
+    const gameStatus = getNewGameStatus(board, currentToken, currentPlayer);
+    return { board, gameStatus };
+};
